Rename packages based on review
