+++ Makefile
@@ -65,8 +65,8 @@ lerna-build: yarn-install
 
 npm-packages: lerna-build
 	mkdir -p dist
-	$(call PACKAGE_LAB_EXTENSION,apputils)
 	$(call PACKAGE_LAB_EXTENSION,application)
+	$(call PACKAGE_LAB_EXTENSION,theme)
 	$(call PACKAGE_LAB_EXTENSION,code-snippet)
 	$(call PACKAGE_LAB_EXTENSION,pipeline-editor)
 	$(call PACKAGE_LAB_EXTENSION,python-runner)
@@ -78,14 +78,14 @@ bdist: npm-packages
 
 install: bdist lint ## Build distribution and install
 	pip install --upgrade dist/elyra-*-py3-none-any.whl
-	$(call UNLINK_LAB_EXTENSION,apputils)
 	$(call UNLINK_LAB_EXTENSION,application)
+	$(call UNLINK_LAB_EXTENSION,theme)
 	$(call UNLINK_LAB_EXTENSION,code-snippet-extension-experimental)
 	$(call UNLINK_LAB_EXTENSION,pipeline-editor-extension)
 	$(call UNLINK_LAB_EXTENSION,python-runner-extension)
 	jupyter lab clean
-	$(call LINK_LAB_EXTENSION,apputils)
 	$(call LINK_LAB_EXTENSION,application)
+	$(call LINK_LAB_EXTENSION,theme)
 	$(call LINK_LAB_EXTENSION,code-snippet)
 	$(call LINK_LAB_EXTENSION,pipeline-editor)
 	$(call LINK_LAB_EXTENSION,python-runner)
