+++ packages/application/src/icons.tsx
