+++ packages/application/src/index.ts
@@ -14,45 +14,9 @@
  * limitations under the License.
  */
 
-import { elyraIcon } from '@elyra/apputils';
-import {
-  JupyterFrontEnd,
-  JupyterFrontEndPlugin
-} from '@jupyterlab/application';
-import { IMainMenu } from '@jupyterlab/mainmenu';
-
 import '../style/index.css';
 
-/**
- * Initialization data for the application extension.
- */
-const extension: JupyterFrontEndPlugin<void> = {
-  id: 'elyra-application',
-  autoStart: true,
-  requires: [IMainMenu],
-  activate: (app: JupyterFrontEnd) => {
-    console.log('Elyra - application extension is activated!');
-
-    // Find the MainLogo widget in the shell and replace it with the Elyra Logo
-    const widgets = app.shell.widgets('top');
-    let widget = widgets.next();
-
-    while (widget !== undefined) {
-      if (widget.id === 'jp-MainLogo') {
-        elyraIcon.element({
-          container: widget.node,
-          justify: 'center',
-          margin: '2px 5px 2px 5px',
-          height: 'auto',
-          width: '20px'
-        });
-        break;
-      }
-
-      widget = widgets.next();
-      console.log(widget);
-    }
-  }
-};
-
-export default extension;
+export * from './parsing';
+export * from './services';
+export * from './submission';
+export * from './icons';
