+++ packages/application/src/svg.d.ts
