+++ packages/application/style/index.css
@@ -13,3 +13,60 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
+.elyra-pipelineSubmission-messageDisplay {
+  padding-bottom: 5px;
+  display: flex;
+  flex-direction: column;
+  height: 100%;
+}
+
+.elyra-pipelineSubmission-errDisplayButton {
+  margin-right: 10px;
+  vertical-align: text-bottom;
+  background-color: white;
+  background-repeat: no-repeat;
+  border: none;
+}
+
+.elyra-pipelineSubmission-errDisplayButton:hover {
+  background-color: var(--jp-layout-color2);
+}
+
+button.elyra-pipelineSubmission-errDisplayButton {
+  height: 20px;
+  width: 20px;
+  padding: 0;
+}
+
+.elyra-pipelineSubmission-errDetails {
+  margin: 15px 0;
+  display: flex;
+  flex: 1;
+  min-height: 0px;
+  flex-direction: column;
+}
+
+.elyra-pipelineSubmission-error-visible {
+  display: block;
+  overflow: auto;
+  border: var(--jp-border-width) solid var(--jp-toolbar-border-color);
+  background-color: #f6f8fa;
+  padding: 5px;
+  flex: 1;
+  overflow: auto;
+  margin: 5px 0 10px 30px;
+}
+
+.elyra-pipelineSubmission-error-hidden {
+  display: none;
+}
+
+/* temporary fix until this is addressed in jupyterlab */
+.lm-TabBar-tabIcon svg {
+  height: auto;
+}
+
+.jp-Dialog-content {
+  resize: both;
+}
