+++ packages/application/tsconfig.json
@@ -4,10 +4,5 @@
     "outDir": "lib",
     "rootDir": "src"
   },
-  "include": ["src/*"],
-  "references": [
-    {
-      "path": "../apputils"
-    }
-  ]
+  "include": ["src/*"]
 }
