+++ packages/code-snippet/tsconfig.json
@@ -7,7 +7,7 @@
   "include": ["src/*"],
   "references": [
     {
-      "path": "../apputils"
+      "path": "../application"
     }
   ]
 }
