+++ packages/pipeline-editor/package.json
@@ -31,7 +31,7 @@
     "watch": "tsc -w"
   },
   "dependencies": {
-    "@elyra/apputils": "^0.11.0-dev",
+    "@elyra/application": "^0.11.0-dev",
     "@elyra/canvas": "6.1.23",
     "@jupyterlab/application": "^2.0.2",
     "@jupyterlab/apputils": "^2.0.2",
