+++ packages/pipeline-editor/src/SubmitNotebook.ts
@@ -18,7 +18,7 @@ import {
   IDictionary,
   NotebookParser,
   SubmissionHandler
-} from '@elyra/apputils';
+} from '@elyra/application';
 import { JupyterFrontEnd } from '@jupyterlab/application';
 import { Dialog, showDialog, ToolbarButton } from '@jupyterlab/apputils';
 import { DocumentRegistry } from '@jupyterlab/docregistry';
