+++ packages/pipeline-editor/src/submission.tsx
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import { SubmissionHandler } from '@elyra/apputils';
+import { SubmissionHandler } from '@elyra/application';
 import { showDialog, Dialog } from '@jupyterlab/apputils';
 import * as React from 'react';
