+++ packages/pipeline-editor/tsconfig.json
@@ -17,7 +17,7 @@
   ],
   "references": [
     {
-      "path": "../apputils"
+      "path": "../application"
     }
   ]
 }
