+++ packages/theme/package.json
@@ -1,10 +1,11 @@
 {
-  "name": "@elyra/apputils",
+  "name": "@elyra/theme",
   "version": "0.11.0-dev",
-  "description": "JupyterLab - Utilities for use in elyra",
+  "description": "JupyterLab extension - Elyra theme",
   "keywords": [
     "jupyter",
-    "jupyterlab"
+    "jupyterlab",
+    "jupyterlab-extension"
   ],
   "homepage": "https://github.com/elyra-ai/elyra",
   "bugs": {
@@ -30,15 +31,16 @@
     "watch": "tsc -w"
   },
   "dependencies": {
+    "@elyra/application": "^0.11.0-dev",
     "@jupyterlab/application": "^2.0.2",
-    "@jupyterlab/apputils": "^2.0.2",
-    "@jupyterlab/coreutils": "^4.0.2",
-    "@jupyterlab/ui-components": "^2.0.2",
-    "@jupyterlab/services": "^5.0.2"
+    "@jupyterlab/mainmenu": "^2.0.2"
   },
   "devDependencies": {
     "rimraf": "~3.0.0",
     "typescript": "~3.7.3"
+  },
+  "jupyterlab": {
+    "extension": true
   },
   "publishConfig": {
     "access": "public"
