+++ packages/theme/style/index.css
@@ -13,10 +13,3 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
-import '../style/index.css';
-
-export * from './parsing';
-export * from './services';
-export * from './submission';
-export * from './icons';
